+++ README.md
@@ -27,9 +27,26 @@ cc.remove_people_profile(website_id: website_id, people_id: person_id)
 
 cc.list_people_profiles(website_id: website_id)
 #=> will return the first 20 users by default
-cc.list_people_profiles(website_id: website_id, page_number: 2)
+cc.list_people_profiles(website_id: website_id, page_number: 2, search_filter: "test@email.com")
 #=> will return the next 20 users
 # Other parameters you can use include: sort_field, sort_order, search_operator, search_filter
+# The search_filter parameter is to be used exactly as in the dashboard, as in: "&search_filter="parameter"
+
+cc.update_conversation_metas website_id: @@website_id, session_id: @@session_id, 
+									meta: { nickname: "test nickname", email: "test@email.com", phone: "9999-9999", 
+													address: "rua teste 99t", segments: ["segmento 1", "segmento 2"], 
+									data: { teste_data: "data_teste_3", valorArbitrario: "teste arbitrário 3" } }
+# Updates the metadata from a specific conversation, given a session_id
+# Remember that conversation metadata is not equal to profile metadata. Each conversation has their own parameters.
+
+cc.update_people_profile website_id: @@website_id, people_id: @@people_id, 
+  								profile_data: { person: { phone: "9999-9999", nickname: "test nickname 2", 
+                                                address: "rua teste 99t", website: "http://teste.com"},
+                                    segments: ["segmento 4", "segmento 3"], 
+                                    company: { name: "Empresa teste" },
+                                    geolocation: { city: "Cidade teste" } }
+# Updates a user's profile data. This is what you see when you click in a specific profile.
+
 ```
 
 ## Installation
