+++ lib/crisp_client.rb
@@ -2,7 +2,7 @@ require "crisp_client/version"
 require "crisp_client/user_availability"
 require "crisp_client/list_websites"
 require "crisp_client/website_people"
-#require "crisp_client/website_conversation"
+require "crisp_client/website_conversation.rb"
 require "httparty"
 require "uri"
 
@@ -13,7 +13,7 @@ module CrispClient
     include ListWebsites
     include WebsitePeople
     include URI
-    #include WebsiteConversation
+    include WebsiteConversation
 
     base_uri "https://api.crisp.im/v1"
