+++ lib/crisp_client/version.rb
@@ -1,3 +1,3 @@
 module CrispClient
-  VERSION = "0.2.4"
+  VERSION = "0.2.5"
 end
