+++ test/integration tests/create_and_update_user_data.rb
@@ -2,9 +2,9 @@ require 'test/unit'
 require 'httparty'
 require 'crisp_client'
 
-
 class TestWebsiteConversation < Test::Unit::TestCase
 	include CrispClient
+
 	def test_authentication
 		assert_not_nil ENV['CRISP_EMAIL'], 'You must set an e-mail as a $CRISP_EMAIL environment variable, ex: CRISP_EMAIL="my@email.com"'
 		assert_not_nil ENV['CRISP_PASSWORD'], 'You must set a password as a $CRISP_PASSWORD environment variable, ex: CRISP_PASSWORD="Mypassword"'
@@ -43,16 +43,14 @@ class TestWebsiteConversation < Test::Unit::TestCase
   	assert_not_nil response, "Current response value: #{response}, person: #{@@people_id}"
   end
 
-#	def test_update_conversation_metas
-#		assert_not_nil @@session_id, "You must provide a session ID. Current session_id value: #{@@session_id}"
-#		
-#		response = @@cclient.update_conversation_metas website_id: @@website_id, session_id: @@session_id, 
-#									meta: { nickname: "test nickname", email: "test@email.com", phone: "9999-9999", 
-#													address: "rua teste 99t", segments: ["segmento 1", "segmento 2"], 
-#									data: { teste_data: "data_teste", valorArbitrario: "teste arbitrário" } }
-#		
-#		assert_not_nil response, "Current response value: #{response}, session_id: #{@@session_id}"
-#	end
+	def test_update_conversation_metas
+		assert_not_nil @@session_id, "You must provide a session ID. Current session_id value: #{@@session_id}"
 
+		response = @@cclient.update_conversation_metas website_id: @@website_id, session_id: @@session_id, 
+									meta: { nickname: "test nickname", email: "test@email.com", phone: "9999-9999", 
+													address: "rua teste 99t", segments: ["segmento 1", "segmento 2"], 
+									data: { teste_data: "data_teste_3", valorArbitrario: "teste arbitrário 3" } }
 
+		assert_not_nil response, "Current response value: #{response}, session_id: #{@@session_id}"
+	end
 end
